Add accordeon widget
The accordeon widget allows the user to insert an array of items and
will display these items in a title/description format. Initially the
description will be hidden, but upon clicking the title the description
will be shown to the user. Only one accordeon can be open at any time.