Change the link field in the icon section to an array
This allows the user to enter as many links as they want per icon step

WARNING: This will break backwards compatibility due to not checking for
the old linkTitle and linkUrl variables. The problem with checking for
those variables is that they are not unsettable anymore. So when we
check for those variables, it's impossible for the user to remove all
links. A possible fix would be to add a question 'Show one or multiple
links?', but this will not make the CMS easier to use. Especially in the
long run.