Fix idea count and allow user to enter fixed idea count in header widget
The idea count in the header widget is now retrieved from the actual
ideas.

There is now a new option to enter a fixed amount of ideas. Only the
first 3 digits entered are used. If the length is less than 3 the amount
is prepended with zeros.

Note: There is no validation currently, so it is allowed to enter text.