Allow combination of BUSY and DENIED status in the overview
This essentially adds a checkbox which - when checked - will change the
BUSY status to DENIED in the overview. Which allows us to use BUSY for
ideas without enough likes, and DENIED for other ideas. We can then
allow the visitor to filter on 'DENIED' and see both these types of
plans as 'DENIED'. We only show the differentiation between these two
statuses on the idea page.