Add check whether attribute is a number or not when sorting
Currently, when an attribute is not available for each element, the
sorting will not work as expected. This happens when we sort by ranking
for example. Not every idea has a ranking attribute, and when we try
to parse this empty data-attribute to an integer, we get a 'NaN', which
will lead to unexpected results in the current sorting check.

By checking with isNaN we have better control over the sorting.