Add functionality to hide ranking
By default the ranking is shown in the participatory budgetting widget, but it's possible to hide the ranking. This allows us to enter the ranking in the database before the actual result has been published. When the result needs to be published, we change the 'showRanking' setting to display the ranking.