Allow accessible navigation through the accordeon widget.html
This turns the title into an 'a'-element so it can be focused and 'clicked' correctly.
After clicking the description div will be focused, so a screen reader will start reading
the content straight away.

Adds a bit of styling to keep the look and feel the same as when the title was a div.