Refactor sendThankYouMail to receive a resourceType
This resourceType is plural, e.g. 'ideas' or 'articles'. To avoid having to pass around an object with a singular and plural version of each resourceType, the {$resourceType}_created.njk files have been renamed to use the plural version.

All calls to the resourceType's config are now routed through the siteConfig object.